requirements.txt
Flask
google-cloud-vision
openai
pandas
python-dotenv
werkzeug
gunicorn